refactor(measurement): split analysis contracts and math
