+++ tools/measurement/analyze_benchmark_output.py
@@ -16,12 +16,89 @@ import json
 import logging
 import math
 import sys
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Annotated, Any, cast
 
 import cyclopts
 import pandas as pd
+from measurement_tools.benchmark_analysis_math import (
+    coalesce_number as _coalesce_number,
+)
+from measurement_tools.benchmark_analysis_math import (
+    f1 as _f1,
+)
+from measurement_tools.benchmark_analysis_math import (
+    first_float as _first_float,
+)
+from measurement_tools.benchmark_analysis_math import (
+    first_int as _first_int,
+)
+from measurement_tools.benchmark_analysis_math import (
+    first_value as _first_value,
+)
+from measurement_tools.benchmark_analysis_math import (
+    model_workflow_rows as _model_workflow_rows,
+)
+from measurement_tools.benchmark_analysis_math import (
+    non_null_count as _non_null_count,
+)
+from measurement_tools.benchmark_analysis_math import (
+    pipeline_stage_rows as _pipeline_stage_rows,
+)
+from measurement_tools.benchmark_analysis_math import (
+    positive_count as _positive_count,
+)
+from measurement_tools.benchmark_analysis_math import (
+    records_of_type as _records_of_type,
+)
+from measurement_tools.benchmark_analysis_math import (
+    request_failure_rate as _request_failure_rate,
+)
+from measurement_tools.benchmark_analysis_math import (
+    safe_rate as _safe_rate,
+)
+from measurement_tools.benchmark_analysis_math import (
+    safe_ratio as _safe_ratio,
+)
+from measurement_tools.benchmark_analysis_math import (
+    sum_bool_or_zero as _sum_bool_or_zero,
+)
+from measurement_tools.benchmark_analysis_math import (
+    sum_int_or_none as _sum_int_or_none,
+)
+from measurement_tools.benchmark_analysis_math import (
+    sum_optional_numbers as _sum_optional_numbers,
+)
+from measurement_tools.benchmark_analysis_math import (
+    sum_prefixed_ints as _sum_prefixed_ints,
+)
+from measurement_tools.benchmark_analysis_math import (
+    zero_count as _zero_count,
+)
+from measurement_tools.benchmark_analysis_math import (
+    zero_with_positive_count as _zero_with_positive_count,
+)
+from measurement_tools.benchmark_analysis_models import (
+    _EVALUATION_ROLLUPS,
+)
+from measurement_tools.benchmark_analysis_models import (
+    BenchmarkOutputAnalysis as BenchmarkOutputAnalysis,
+)
+from measurement_tools.benchmark_analysis_models import (
+    CaseAnalysisRow as CaseAnalysisRow,
+)
+from measurement_tools.benchmark_analysis_models import (
+    GroupAnalysisRow as GroupAnalysisRow,
+)
+from measurement_tools.benchmark_analysis_models import (
+    ModelUsageAnalysisRow as ModelUsageAnalysisRow,
+)
+from measurement_tools.benchmark_analysis_models import (
+    ModelUsageGroupAnalysisRow as ModelUsageGroupAnalysisRow,
+)
+from measurement_tools.benchmark_analysis_models import (
+    _EvaluationRollup as _EvaluationRollup,
+)
 from measurement_tools.cli import LogFormat, configure_logging, log_bad_input
 from measurement_tools.stats import median_or_none as _median_or_none
 from measurement_tools.stats import none_if_nan as _none_if_nan
@@ -30,7 +107,6 @@ from measurement_tools.stats import sum_or_none as _sum_or_none
 from measurement_tools.stats import sum_or_zero as _sum_or_zero
 from measurement_tools.tables import AnalysisExportResult, ExportFormat, ModelTableSpec
 from measurement_tools.tables import write_analysis_tables as _write_analysis_table_specs
-from pydantic import BaseModel, Field, computed_field
 
 app = cyclopts.App(help=__doc__)
 logger = logging.getLogger("measurement.benchmark_output")
@@ -44,319 +120,6 @@ _SIGNATURE_DETAIL_FIELDS = {
     "end_position",
     "value_length",
 }
-
-
-@dataclass(frozen=True)
-class _EvaluationRollup:
-    prefix: str
-    valid_column: str
-    invalid_count_column: str
-
-
-_EVALUATION_ROLLUPS = (
-    _EvaluationRollup("detection", "detection_valid", "detection_invalid_entity_count"),
-    _EvaluationRollup("type_fidelity", "type_fidelity_valid", "type_fidelity_invalid_replacement_count"),
-    _EvaluationRollup(
-        "relational_consistency",
-        "relational_consistency_valid",
-        "relational_consistency_invalid_relation_count",
-    ),
-    _EvaluationRollup("attribute_fidelity", "attribute_fidelity_valid", "attribute_fidelity_invalid_entity_count"),
-)
-
-
-class CaseAnalysisRow(BaseModel):
-    suite_id: str | None = None
-    workload_id: str | None = None
-    workload_category: str | None = None
-    config_id: str | None = None
-    experimental_detection_strategy: str | None = None
-    experimental_replacement_strategy: str | None = None
-    dd_parser_compat: str | None = None
-    entity_label_set_id: str | None = None
-    entity_label_count: int | None = None
-    gliner_threshold: float | None = None
-    repetition: int | None = None
-    case_id: str
-    run_id: str
-    case_failed: bool = False
-    error_stage_count: int = 0
-    error_ndd_workflow_count: int = 0
-    error_model_workflow_count: int = 0
-    pipeline_elapsed_sec: float | None = None
-    ndd_workflow_count: int = 0
-    ndd_elapsed_sec_total: float = 0.0
-    observed_total_requests: int = 0
-    observed_successful_requests: int = 0
-    observed_input_tokens: int = 0
-    observed_output_tokens: int = 0
-    observed_total_tokens: int = 0
-    observed_failed_requests: int = 0
-    observed_failed_request_rate: float | None = None
-    dd_trace_record_count: int = 0
-    dd_trace_error_count: int = 0
-    dd_trace_sync_client_unavailable_count: int = 0
-    observed_bridge_fallback_requests: int | None = None
-    observed_non_bridge_total_requests: int | None = None
-    observed_non_bridge_failed_requests: int | None = None
-    observed_non_bridge_failed_request_rate: float | None = None
-    record_count: int = 0
-    input_text_tokens_total: int | None = None
-    records_per_pipeline_sec: float | None = None
-    records_per_ndd_sec: float | None = None
-    input_text_tokens_per_pipeline_sec: float | None = None
-    input_text_tokens_per_ndd_sec: float | None = None
-    topology_endpoint_count: float | None = None
-    topology_gpu_count: float | None = None
-    topology_tensor_parallelism: float | None = None
-    topology_shard_count: float | None = None
-    input_text_tokens_per_endpoint_sec: float | None = None
-    input_text_tokens_per_gpu_sec: float | None = None
-    final_entity_count: float | None = None
-    empty_detection_count: int = 0
-    empty_detection_rate: float | None = None
-    empty_detection_with_ground_truth_count: int = 0
-    empty_detection_with_ground_truth_rate: float | None = None
-    ground_truth_record_count: int = 0
-    ground_truth_entity_count: float | None = None
-    entity_true_positive_count: float | None = None
-    entity_false_positive_count: float | None = None
-    entity_false_negative_count: float | None = None
-    entity_precision: float | None = None
-    entity_recall: float | None = None
-    entity_f1: float | None = None
-    entity_relaxed_gt_found_count: float | None = None
-    entity_relaxed_detected_tp_count: float | None = None
-    entity_relaxed_label_compatible_gt_found_count: float | None = None
-    entity_relaxed_label_compatible_detected_tp_count: float | None = None
-    entity_relaxed_precision: float | None = None
-    entity_relaxed_recall: float | None = None
-    entity_relaxed_f1: float | None = None
-    entity_relaxed_label_compatible_precision: float | None = None
-    entity_relaxed_label_compatible_recall: float | None = None
-    entity_relaxed_label_compatible_f1: float | None = None
-    replacement_count: float | None = None
-    replacement_missing_final_entity_count: float | None = None
-    replacement_missing_final_entity_label_counts: dict[str, int] = Field(default_factory=dict)
-    replacement_missing_final_value_count: float | None = None
-    replacement_synthetic_original_collision_count: float | None = None
-    replacement_synthetic_original_collision_label_counts: dict[str, int] = Field(default_factory=dict)
-    replacement_synthetic_original_collision_value_count: float | None = None
-    original_value_leak_count: float | None = None
-    original_value_leak_record_count: int = 0
-    original_value_leak_label_counts: dict[str, int] = Field(default_factory=dict)
-    detection_judged_record_count: int = 0
-    detection_valid_record_count: int = 0
-    detection_valid_rate: float | None = None
-    detection_invalid_entity_count: int = 0
-    type_fidelity_judged_record_count: int = 0
-    type_fidelity_valid_record_count: int = 0
-    type_fidelity_valid_rate: float | None = None
-    type_fidelity_invalid_replacement_count: int = 0
-    relational_consistency_judged_record_count: int = 0
-    relational_consistency_valid_record_count: int = 0
-    relational_consistency_valid_rate: float | None = None
-    relational_consistency_invalid_relation_count: int = 0
-    attribute_fidelity_judged_record_count: int = 0
-    attribute_fidelity_valid_record_count: int = 0
-    attribute_fidelity_valid_rate: float | None = None
-    attribute_fidelity_invalid_entity_count: int = 0
-    validation_max_entities_per_call: int | None = None
-    detection_artifact_rows: int = 0
-    seed_entity_count: float | None = None
-    seed_validation_candidate_count: float | None = None
-    estimated_seed_validation_chunk_count: float | None = None
-    augmented_entity_count: float | None = None
-    augmented_new_final_value_count: float | None = None
-    artifact_final_entity_count: float | None = None
-    artifact_final_detector_entity_count: float | None = None
-    artifact_final_augmenter_entity_count: float | None = None
-    artifact_final_entity_signature_count: float | None = None
-    artifact_final_entity_signature_hashes: list[str] = Field(default_factory=list)
-    artifact_final_entity_signature_labels: dict[str, str] = Field(default_factory=dict)
-    artifact_final_entity_signature_details: dict[str, dict[str, Any]] = Field(default_factory=dict)
-
-
-class GroupAnalysisRow(BaseModel):
-    workload_id: str | None = None
-    workload_category: str | None = None
-    config_id: str | None = None
-    experimental_detection_strategy: str | None = None
-    experimental_replacement_strategy: str | None = None
-    entity_label_set_id: str | None = None
-    entity_label_count: int | None = None
-    gliner_threshold: float | None = None
-    case_count: int
-    failed_case_count: int = 0
-    failed_case_rate: float | None = None
-    error_stage_count: int = 0
-    error_ndd_workflow_count: int = 0
-    error_model_workflow_count: int = 0
-    median_pipeline_elapsed_sec: float | None = None
-    median_ndd_elapsed_sec_total: float | None = None
-    median_observed_total_requests: float | None = None
-    median_observed_successful_requests: float | None = None
-    median_observed_input_tokens: float | None = None
-    median_observed_output_tokens: float | None = None
-    median_observed_total_tokens: float | None = None
-    median_observed_failed_requests: float | None = None
-    median_observed_failed_request_rate: float | None = None
-    median_observed_bridge_fallback_requests: float | None = None
-    median_observed_non_bridge_total_requests: float | None = None
-    median_observed_non_bridge_failed_requests: float | None = None
-    median_observed_non_bridge_failed_request_rate: float | None = None
-    total_record_count: int = 0
-    median_record_count: float | None = None
-    total_input_text_tokens: int | None = None
-    median_input_text_tokens_total: float | None = None
-    median_records_per_pipeline_sec: float | None = None
-    median_records_per_ndd_sec: float | None = None
-    median_input_text_tokens_per_pipeline_sec: float | None = None
-    median_input_text_tokens_per_ndd_sec: float | None = None
-    median_topology_endpoint_count: float | None = None
-    median_topology_gpu_count: float | None = None
-    median_topology_tensor_parallelism: float | None = None
-    median_topology_shard_count: float | None = None
-    median_input_text_tokens_per_endpoint_sec: float | None = None
-    median_input_text_tokens_per_gpu_sec: float | None = None
-    median_final_entity_count: float | None = None
-    total_empty_detection_count: int = 0
-    empty_detection_rate: float | None = None
-    total_empty_detection_with_ground_truth_count: int = 0
-    empty_detection_with_ground_truth_rate: float | None = None
-    total_ground_truth_record_count: int = 0
-    sum_ground_truth_entity_count: float | None = None
-    sum_entity_true_positive_count: float | None = None
-    sum_entity_false_positive_count: float | None = None
-    sum_entity_false_negative_count: float | None = None
-    micro_entity_precision: float | None = None
-    micro_entity_recall: float | None = None
-    micro_entity_f1: float | None = None
-    sum_entity_relaxed_gt_found_count: float | None = None
-    sum_entity_relaxed_detected_tp_count: float | None = None
-    sum_entity_relaxed_label_compatible_gt_found_count: float | None = None
-    sum_entity_relaxed_label_compatible_detected_tp_count: float | None = None
-    micro_entity_relaxed_precision: float | None = None
-    micro_entity_relaxed_recall: float | None = None
-    micro_entity_relaxed_f1: float | None = None
-    micro_entity_relaxed_label_compatible_precision: float | None = None
-    micro_entity_relaxed_label_compatible_recall: float | None = None
-    micro_entity_relaxed_label_compatible_f1: float | None = None
-    median_entity_relaxed_f1: float | None = None
-    median_entity_relaxed_label_compatible_f1: float | None = None
-    median_replacement_missing_final_entity_count: float | None = None
-    median_replacement_missing_final_value_count: float | None = None
-    replacement_missing_final_entity_label_counts: dict[str, int] = Field(default_factory=dict)
-    median_replacement_synthetic_original_collision_count: float | None = None
-    median_replacement_synthetic_original_collision_value_count: float | None = None
-    replacement_synthetic_original_collision_label_counts: dict[str, int] = Field(default_factory=dict)
-    sum_original_value_leak_count: float | None = None
-    leaking_case_count: int = 0
-    median_original_value_leak_count: float | None = None
-    sum_detection_judged_record_count: int = 0
-    sum_detection_valid_record_count: int = 0
-    micro_detection_valid_rate: float | None = None
-    sum_detection_invalid_entity_count: int = 0
-    sum_type_fidelity_judged_record_count: int = 0
-    sum_type_fidelity_valid_record_count: int = 0
-    micro_type_fidelity_valid_rate: float | None = None
-    sum_type_fidelity_invalid_replacement_count: int = 0
-    sum_relational_consistency_judged_record_count: int = 0
-    sum_relational_consistency_valid_record_count: int = 0
-    micro_relational_consistency_valid_rate: float | None = None
-    sum_relational_consistency_invalid_relation_count: int = 0
-    sum_attribute_fidelity_judged_record_count: int = 0
-    sum_attribute_fidelity_valid_record_count: int = 0
-    micro_attribute_fidelity_valid_rate: float | None = None
-    sum_attribute_fidelity_invalid_entity_count: int = 0
-    median_seed_entity_count: float | None = None
-    median_seed_validation_candidate_count: float | None = None
-    median_estimated_seed_validation_chunk_count: float | None = None
-    median_augmented_entity_count: float | None = None
-    median_augmented_new_final_value_count: float | None = None
-    median_artifact_final_entity_count: float | None = None
-    median_artifact_final_detector_entity_count: float | None = None
-    median_artifact_final_augmenter_entity_count: float | None = None
-    median_artifact_final_entity_signature_count: float | None = None
-
-
-class ModelUsageAnalysisRow(BaseModel):
-    suite_id: str | None = None
-    workload_id: str | None = None
-    config_id: str | None = None
-    experimental_detection_strategy: str | None = None
-    experimental_replacement_strategy: str | None = None
-    dd_parser_compat: str | None = None
-    repetition: int | None = None
-    case_id: str
-    run_id: str
-    workflow_name: str | None = None
-    model_alias: str | None = None
-    model_name: str
-    model_provider_name: str | None = None
-    ndd_elapsed_sec: float | None = None
-    observed_total_requests: int = 0
-    observed_successful_requests: int = 0
-    observed_failed_requests: int = 0
-    observed_input_tokens: int = 0
-    observed_output_tokens: int = 0
-    observed_total_tokens: int = 0
-    observed_reasoning_tokens: int | None = None
-    observed_failed_request_rate: float | None = None
-
-
-class ModelUsageGroupAnalysisRow(BaseModel):
-    workload_id: str | None = None
-    config_id: str | None = None
-    experimental_detection_strategy: str | None = None
-    experimental_replacement_strategy: str | None = None
-    dd_parser_compat: str | None = None
-    workflow_name: str | None = None
-    model_alias: str | None = None
-    model_name: str
-    model_provider_name: str | None = None
-    case_count: int
-    workflow_count: int
-    sum_observed_total_requests: int = 0
-    sum_observed_successful_requests: int = 0
-    sum_observed_failed_requests: int = 0
-    sum_observed_input_tokens: int = 0
-    sum_observed_output_tokens: int = 0
-    sum_observed_total_tokens: int = 0
-    sum_observed_reasoning_tokens: int | None = None
-    observed_failed_request_rate: float | None = None
-    median_observed_total_requests: float | None = None
-    median_observed_failed_requests: float | None = None
-    median_observed_total_tokens: float | None = None
-
-
-class BenchmarkOutputAnalysis(BaseModel):
-    benchmark_dir: str
-    detection_artifacts_path: str | None = None
-    cases: list[CaseAnalysisRow] = Field(default_factory=list)
-    groups: list[GroupAnalysisRow] = Field(default_factory=list)
-    model_usage: list[ModelUsageAnalysisRow] = Field(default_factory=list)
-    model_usage_groups: list[ModelUsageGroupAnalysisRow] = Field(default_factory=list)
-
-    @computed_field
-    @property
-    def case_count(self) -> int:
-        return len(self.cases)
-
-    @computed_field
-    @property
-    def group_count(self) -> int:
-        return len(self.groups)
-
-    @computed_field
-    @property
-    def model_usage_count(self) -> int:
-        return len(self.model_usage)
-
-    @computed_field
-    @property
-    def model_usage_group_count(self) -> int:
-        return len(self.model_usage_groups)
 
 
 def analyze_benchmark_output(
@@ -823,29 +586,6 @@ def _rows_for_case(dataframe: pd.DataFrame, case_id: str) -> pd.DataFrame:
     return dataframe[mask]
 
 
-def _records_of_type(dataframe: pd.DataFrame, record_type: str) -> pd.DataFrame:
-    if "record_type" not in dataframe.columns:
-        return dataframe.iloc[0:0]
-    return dataframe[dataframe["record_type"] == record_type]
-
-
-def _records_of_types(dataframe: pd.DataFrame, record_types: set[str]) -> pd.DataFrame:
-    if "record_type" not in dataframe.columns:
-        return dataframe.iloc[0:0]
-    return dataframe[dataframe["record_type"].isin(record_types)]
-
-
-def _model_workflow_rows(dataframe: pd.DataFrame) -> pd.DataFrame:
-    return _records_of_types(dataframe, {"ndd_workflow", "model_workflow"})
-
-
-def _pipeline_stage_rows(dataframe: pd.DataFrame) -> pd.DataFrame:
-    stages = _records_of_type(dataframe, "stage")
-    if "stage" not in stages.columns:
-        return stages.iloc[0:0]
-    return stages[stages["stage"] == "Anonymizer._run_internal"]
-
-
 _MODEL_USAGE_SUFFIXES = {
     ".request_usage.total_requests": "observed_total_requests",
     ".request_usage.successful_requests": "observed_successful_requests",
@@ -988,34 +728,6 @@ def _coerce_int(value: Any) -> int:
     return int(float(value))
 
 
-def _first_value(frames: list[pd.DataFrame], columns: list[str]) -> str | None:
-    for frame in frames:
-        for column in columns:
-            if column not in frame.columns:
-                continue
-            values = frame[column].dropna()
-            if not values.empty:
-                return str(values.iloc[0])
-    return None
-
-
-def _first_int(frames: list[pd.DataFrame], columns: list[str]) -> int | None:
-    value = _first_value(frames, columns)
-    return int(float(value)) if value is not None else None
-
-
-def _first_float(frames: list[pd.DataFrame], columns: list[str]) -> float | None:
-    value = _first_value(frames, columns)
-    return float(value) if value is not None else None
-
-
-def _coalesce_number(*values: float | None) -> float | None:
-    for value in values:
-        if value is not None:
-            return value
-    return None
-
-
 def _artifact_signature_hashes(artifact_rows: pd.DataFrame) -> list[str]:
     if "final_entity_signature_hashes" not in artifact_rows.columns:
         return []
@@ -1102,65 +814,6 @@ def _signature_count(artifact_rows: pd.DataFrame, *, signature_hashes: list[str]
     if signature_hashes:
         return float(len(signature_hashes))
     return _sum_or_none(artifact_rows, "final_entity_signature_count")
-
-
-def _positive_count(dataframe: pd.DataFrame, column: str) -> int:
-    if column not in dataframe.columns:
-        return 0
-    values = pd.to_numeric(dataframe[column], errors="coerce").fillna(0)
-    return int((values > 0).sum())
-
-
-def _zero_count(dataframe: pd.DataFrame, column: str) -> int:
-    if column not in dataframe.columns:
-        return 0
-    values = pd.to_numeric(dataframe[column], errors="coerce").dropna()
-    return int((values == 0).sum())
-
-
-def _non_null_count(dataframe: pd.DataFrame, column: str) -> int:
-    if column not in dataframe.columns:
-        return 0
-    return int(pd.to_numeric(dataframe[column], errors="coerce").notna().sum())
-
-
-def _zero_with_positive_count(dataframe: pd.DataFrame, *, zero_column: str, positive_column: str) -> int:
-    if zero_column not in dataframe.columns or positive_column not in dataframe.columns:
-        return 0
-    zero_values = pd.to_numeric(dataframe[zero_column], errors="coerce")
-    positive_values = pd.to_numeric(dataframe[positive_column], errors="coerce")
-    return int(((zero_values == 0) & (positive_values > 0)).sum())
-
-
-def _sum_prefixed_ints(dataframe: pd.DataFrame, prefix: str) -> dict[str, int]:
-    totals: dict[str, int] = {}
-    base_column = prefix.removesuffix(".")
-    if base_column in dataframe.columns:
-        for value in dataframe[base_column].dropna().tolist():
-            for key, count in _coerce_count_mapping(value).items():
-                totals[key] = totals.get(key, 0) + count
-    for column in sorted(col for col in dataframe.columns if col.startswith(prefix)):
-        value = _sum_int_or_zero(dataframe, column)
-        if value:
-            totals[column.removeprefix(prefix)] = value
-    return totals
-
-
-def _coerce_count_mapping(value: object) -> dict[str, int]:
-    payload = value
-    if isinstance(value, str):
-        try:
-            payload = json.loads(value)
-        except json.JSONDecodeError:
-            return {}
-    if not isinstance(payload, dict):
-        return {}
-    counts: dict[str, int] = {}
-    for key, count in payload.items():
-        numeric = pd.to_numeric(pd.Series([count]), errors="coerce").dropna()
-        if not numeric.empty and numeric.iloc[0]:
-            counts[str(key)] = int(numeric.iloc[0])
-    return counts
 
 
 def build_group_rows(cases: list[CaseAnalysisRow]) -> list[GroupAnalysisRow]:
@@ -1404,12 +1057,6 @@ def _float_if_not_nan(value: object) -> float | None:
     return float(cast(Any, value))
 
 
-def _sum_bool_or_zero(dataframe: pd.DataFrame, column: str) -> int:
-    if column not in dataframe.columns:
-        return 0
-    return int(dataframe[column].fillna(False).astype(bool).sum())
-
-
 def _group_evaluation_metrics(group: pd.DataFrame) -> dict[str, int | float | None]:
     metrics: dict[str, int | float | None] = {}
     for rollup in _EVALUATION_ROLLUPS:
@@ -1420,54 +1067,6 @@ def _group_evaluation_metrics(group: pd.DataFrame) -> dict[str, int | float | No
         metrics[f"micro_{rollup.prefix}_valid_rate"] = _safe_ratio(valid_count, judged_count)
         metrics[f"sum_{rollup.invalid_count_column}"] = _sum_int_or_zero(group, rollup.invalid_count_column)
     return metrics
-
-
-def _sum_int_or_none(dataframe: pd.DataFrame, column: str) -> int | None:
-    value = _sum_or_none(dataframe, column)
-    return int(value) if value is not None else None
-
-
-def _request_failure_rate(*, failed: object, total: object) -> float | None:
-    failed_value = _optional_number(failed)
-    total_value = _optional_number(total)
-    if failed_value is None or total_value is None or total_value <= 0:
-        return None
-    return failed_value / total_value
-
-
-def _safe_rate(numerator: object, elapsed_sec: object) -> float | None:
-    numerator_value = _optional_number(numerator)
-    elapsed_value = _optional_number(elapsed_sec)
-    if numerator_value is None or elapsed_value is None or elapsed_value <= 0:
-        return None
-    return numerator_value / elapsed_value
-
-
-def _safe_ratio(numerator: object, denominator: object) -> float | None:
-    numerator_value = _optional_number(numerator)
-    denominator_value = _optional_number(denominator)
-    if numerator_value is None or denominator_value is None or denominator_value <= 0:
-        return None
-    return numerator_value / denominator_value
-
-
-def _sum_optional_numbers(*values: object) -> float | None:
-    numeric_values = [_optional_number(value) for value in values]
-    if any(value is None for value in numeric_values):
-        return None
-    return sum(cast(float, value) for value in numeric_values)
-
-
-def _f1(precision: float | None, recall: float | None) -> float | None:
-    if precision is None or recall is None or precision + recall == 0:
-        return None
-    return 2 * precision * recall / (precision + recall)
-
-
-def _optional_number(value: object) -> float | None:
-    if value is None or pd.isna(value):
-        return None
-    return float(value)
 
 
 def _estimated_validation_chunk_count(
